+++ lib/Castle.php
@@ -21,7 +21,6 @@ require(dirname(__FILE__) . '/Castle/Errors.php');
 require(dirname(__FILE__) . '/RestModel/Resource.php');
 require(dirname(__FILE__) . '/RestModel/Model.php');
 require(dirname(__FILE__) . '/Castle/Context.php');
-require(dirname(__FILE__) . '/Castle/Authenticate.php');
 require(dirname(__FILE__) . '/Castle/CurlTransport.php');
 require(dirname(__FILE__) . '/Castle/RequestContext.php');
 require(dirname(__FILE__) . '/Castle/Request.php');