+++ lib/Castle/Authenticate.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Castle;
-
-class Authenticate extends RestModel
-{
-  protected $isSingular = true;
-}